+++ app/build.gradle
@@ -27,8 +27,7 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
     buildFeatures {
-        //viewBinding true
-        dataBinding true
+        viewBinding true
     }
 }
 
@@ -45,5 +44,4 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
 }
