+++ build.gradle
@@ -1,25 +1,5 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-buildscript {
-    repositories {
-        maven {url "https://jitpack.io"}
-        google()
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:7.4.2'
-        //implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
-    }
-}
-
 plugins {
     id 'com.android.application' version '7.4.2' apply false
     id 'com.android.library' version '7.4.2' apply false
-}
-
-allprojects {
-    repositories {
-        maven { url 'https://jitpack.io' }
-        google()
-        jcenter()
-    }
 }
